+++ examples/provider/provider.tf
@@ -1,3 +1,12 @@
-provider "scaffolding" {
-  # example configuration here
+terraform {
+  required_providers {
+    letscloud = {
+      source = "letscloud-community/letscloud"
+    }
+  }
+}
+
+provider "letscloud" {
+  # Configure the LetsCloud Provider
+  api_token = "your-api-token" # or use LETSCLOUD_API_KEY env variable
 }